refactor: 이미지 파싱 url 수정
