+++ frontend/src/pages/guest/workDetail/GuestWorkDetail.tsx
@@ -6,7 +6,7 @@ import Footer from '../../../components/@common/footer/Footer';
 import { useToast } from '../../../hooks/@common/useToast';
 import { DividerLine } from '../../../styles/@common/DividerLine.styles';
 import type { WorkDetail } from '../../../types/domain/work.type';
-import { buildThumbnailUrl } from '../../../utils/imageUrl';
+import { buildThumbnailUrl } from '../../../utils/buildThumbnailUrl';
 import * as C from '../../WorkDetail.common.styles';
 import * as S from './GuestWorkDetail.styles';
 
@@ -69,7 +69,7 @@ const GuestWorkDetail = () => {
         {photos.map((photo, index) => (
           <C.ImageContainer
             key={photo.id}
-            src={buildThumbnailUrl(spaceCode, photo.path, '800')}
+            src={buildThumbnailUrl(photo.path, '800')}
             alt={`work-detail-${index}`}
           />
         ))}
